+++ backend/canisters/community/impl/src/updates/add_reaction.rs
@@ -78,7 +78,7 @@ fn add_reaction_impl(args: Args, state: &mut RuntimeState) -> Response {
                                         thread_root_message_index: args.thread_root_message_index,
                                         message_index: message.message_index,
                                         message_event_index: event_index,
-                                        community_name: state.data.name.clone(),
+                                        community_name: state.data.name.value.clone(),
                                         channel_name: channel.chat.name.value.clone(),
                                         added_by: user_id,
                                         added_by_name: args.username,